+++ audit/logging_for_customers.csv
@@ -1,1 +1,29 @@
 ACTION,TIMESTAMP
+LastName1 FirstName1 received 1000.0RON in the current account with IBAN RO09FBOR77971890051360302664707462,05-02-2022 21:52:43
+LastName2 FirstName2 received 1000.0RON in the current account with IBAN RO09FBOR99630222573054385015767051,05-02-2022 21:52:43
+LastName3 FirstName3 received 1000.0RON in the current account with IBAN RO09FBOR95564592353752061963174250,05-02-2022 21:52:43
+LastName4 FirstName4 received 1000.0RON in the current account with IBAN RO09FBOR43248908097809710701388449,05-02-2022 21:52:43
+SC Software integration SRL received 1000.0RON in the current account with IBAN RO09FBOR14604251942325876083032505,05-02-2022 21:52:43
+SC Software development SRL received 1000.0RON in the current account with IBAN RO09FBOR29010589048258359095135357,05-02-2022 21:52:43
+SC Code Masters  SA received 1000.0RON in the current account with IBAN RO09FBOR76585660825878442506230487,05-02-2022 21:52:43
+SC Coding Anything  SA received 1000.0RON in the current account with IBAN RO09FBOR54545018108522602821911997,05-02-2022 21:52:43
+LastName1 FirstName1 logged in,07-01-2022 22:14:46
+LastName1 FirstName1 logged in,07-01-2022 22:15:46
+LastName1 FirstName1 viewed its products,07-01-2022 22:15:59
+LastName1 FirstName1 viewed its current accounts,07-01-2022 22:16:04
+LastName1 FirstName1 viewed its debit cards,07-01-2022 22:16:27
+LastName1 FirstName1 viewed its deposits,07-01-2022 22:16:32
+LastName1 FirstName1 viewed its loans,07-01-2022 22:17:20
+LastName1 FirstName1 logged in,09-10-2022 22:20:06
+LastName1 FirstName1 viewed its deposits,09-10-2022 22:20:12
+LastName1 FirstName1 viewed the available interest rates for deposit and loans,09-10-2022 22:20:29
+LastName1 FirstName1 viewed the exchange rates,09-10-2022 22:20:30
+LastName1 FirstName1 created a new current account in CHF with IBAN RO09FBOR64062931925950907483640704,09-10-2022 22:21:14
+LastName1 FirstName1 viewed its current accounts,09-10-2022 22:21:47
+LastName1 FirstName1 generated a statement for the current account with IBAN RO09FBOR77971890051360302664707462,09-10-2022 22:22:37
+LastName1 FirstName1 viewed its loans,09-10-2022 22:23:58
+LastName1 FirstName1 exchanged 100.0RON for CHF,09-10-2022 22:25:14
+LastName1 FirstName1 logged in,09-10-2022 22:46:24
+LastName1 FirstName1 viewed its current accounts,09-10-2022 22:46:35
+LastName1 FirstName1 generated a statement for the current account with IBAN RO09FBOR64062931925950907483640704,09-10-2022 22:46:47
+LastName1 FirstName1 generated a statement for the current account with IBAN RO09FBOR77971890051360302664707462,09-10-2022 22:46:57
