+++ data_storage/current_account.csv
@@ -1,11 +1,12 @@
 IBAN,AMOUNT,CURRENCY_CODE,OPENING_DATE,CUSTOMER_ID
-RO09FBOR77971890051360302664707462,5500.0,RON,2022-05-02,1930729000000
+RO09FBOR77971890051360302664707462,8065.625,RON,2022-05-02,1930729000000
 RO09FBOR32815446179549736350109087,0.0,USD,2022-05-02,1930729000000
-RO09FBOR99630222573054385015767051,0.0,RON,2022-05-02,1910422111111
-RO09FBOR95564592353752061963174250,0.0,RON,2022-05-02,2880304222222
-RO09FBOR43248908097809710701388449,0.0,RON,2022-05-02,2950715333333
-RO09FBOR14604251942325876083032505,0.0,RON,2022-05-02,111111
-RO09FBOR29010589048258359095135357,0.0,RON,2022-05-02,222222
-RO09FBOR76585660825878442506230487,0.0,RON,2022-05-02,333333
-RO09FBOR54545018108522602821911997,0.0,RON,2022-05-02,444444
+RO09FBOR64062931925950907483640704,20.263424518743665,CHF,2022-09-10,1930729000000
+RO09FBOR54545018108522602821911997,1000.0,RON,2022-05-02,444444
 RO09FBOR00955688336828436626623352,0.0,EUR,2022-05-02,444444
+RO09FBOR99630222573054385015767051,1000.0,RON,2022-05-02,1910422111111
+RO09FBOR95564592353752061963174250,1000.0,RON,2022-05-02,2880304222222
+RO09FBOR43248908097809710701388449,1000.0,RON,2022-05-02,2950715333333
+RO09FBOR14604251942325876083032505,1000.0,RON,2022-05-02,111111
+RO09FBOR29010589048258359095135357,1000.0,RON,2022-05-02,222222
+RO09FBOR76585660825878442506230487,1000.0,RON,2022-05-02,333333
