+++ data_storage/customers.csv
@@ -1,9 +1,9 @@
 CUSTOMER_TYPE,CUSTOMER_ID,CUSTOMER_NAME,BIRTH_DATE,HASH_OF_PASSWORD,PHONE_NUMBER,EMAIL_ADDRESS,ADDRESS_COUNTRY,ADDRESS_CITY,ADDRESS_ZIP_CODE,ADDRESS_STREET_NAME,ADDRESS_STREET_NUMBER,ADDRESS_ADDITIONAL_INFO
 INDIVIDUAL,1930729000000,LastName1 FirstName1,1993-07-29,65bb1cc2fd5feddea98e1d9e3ec89fae2a3f50ed42d6fb96962ef8cd0e2cfcb7,0799999999,firstname1.lastname1@mailserver.com.,Romania,Bucharest,060542,Mehadia,12,NA
+COMPANY,444444,SC Coding Anything  SA,2015-10-22,8952678a398bf6ed1a5e9b333dc055b5dc74a3f27b48485da116ce43a433470c,0788888888,contact@codinganything.ro,Romania,Bucharest,050044,Piata Unirii,15,NA
 INDIVIDUAL,1910422111111,LastName2 FirstName2,1991-04-22,102c304fa0934097958dfb45fec492d8a4f1610c33211944a5a1eeefa4c93de5,0700000000,firstname2.lastname2@mailserver.com,Romania,Bucharest,221122,Ceahlau,14,NA
 INDIVIDUAL,2880304222222,LastName3 FirstName3,1988-03-04,8e1f27aaf4f508836284575560431e21f1990ee35909c0233c747d8ba82e82b8,0711111111,firstname3.lastname3@mailserver.com,Romania,Bucharest,331144,Moxa,8,NA
 INDIVIDUAL,2950715333333,LastName4 FirstName4,1995-07-15,b44e8044471ae1564344c0f34c01d09a99de902f758bd3a600b17212d3fbb66f,0722222222,firstname4.lastname4@mailserver.com,Romania,Piatra-Neamt,775577,Bistritei,3,NA
 COMPANY,111111,SC Software integration SRL,2010-10-10,bcd1a821518b6882809e7e4298cbf6a93fc1f5c0a86686aba435acdf3e0e8be4,0733333333,contact@softwareintegration.ro,Romania,Timisoara,557755,Bega,2,NA
 COMPANY,222222,SC Software development SRL,2005-05-15,a9e7141b897c1d5e6b3446f5146f8e5ad919a4fe9fdcc5f9e0a1318ce32f67e1,0744444444,contact@softwaredevelopment.ro,Romania,Brasov,114411,Turnului,10,NA
 COMPANY,333333,SC Code Masters  SA,2019-10-15,2b0e87919dddfff2762decc1aba687dc3d811644da625b44b30feca0f25d7d03,0755555555,contact@codemasters.ro,Romania,Bucharest,440011,Calea Victoriei,100,NA
-COMPANY,444444,SC Coding Anything  SA,2015-10-22,8952678a398bf6ed1a5e9b333dc055b5dc74a3f27b48485da116ce43a433470c,0788888888,contact@codinganything.ro,Romania,Bucharest,050044,Piata Unirii,15,NA
