+++ data_storage/deposit.csv
@@ -1,3 +1,2 @@
 DEPOSIT_ID,DEPOSIT_AMOUNT,INTEREST_RATE,INTEREST_AMOUNT,DEPOSIT_MATURITY,ASSOCIATED_IBAN,CURRENCY_CODE,OPENING_DATE,CUSTOMER_ID
 00237347827348897501,5000.0,3.2,160.0,2023-05-02,RO09FBOR77971890051360302664707462,RON,2022-05-02,1930729000000
-93131722817163834710,2000.0,2.5,12.5,2022-08-02,RO09FBOR77971890051360302664707462,RON,2022-05-02,1930729000000
