+++ data_storage/loan.csv
@@ -1,2 +1,2 @@
 LOAN_ID,INTEREST_RATE,MATURITY_IN_MONTHS,INDEX_TO_NEXT_PAYMENT,LOAN_INITIAL_AMOUNT,LOAN_CURRENT_AMOUNT,ASSOCIATED_IBAN,CURRENCY_CODE,OPENING_DATE,CUSTOMER_ID
-88969347537438887229,10.0,48,0,3000.0,3000.0,RO09FBOR77971890051360302664707462,RON,2022-05-02,1930729000000
+88969347537438887229,10.0,48,4,3000.0,2750.0,RO09FBOR77971890051360302664707462,RON,2022-05-02,1930729000000
