+++ data_storage/static_variables.csv
@@ -1,6 +1,6 @@
 CLASS_NAME,COUNT_OF
 customer,8
 deposit,2
-current_account,10
+current_account,11
 debit_card,1
 loan,1
