+++ data_storage/system_date.csv
@@ -1,2 +1,2 @@
 SYSTEM_DATE
-2022-05-02
+2022-09-10
