+++ data_storage/transaction_logger.csv
@@ -5,3 +5,23 @@ bf068c81-e235-4ca7-9b93-a7f320dc1963,2022-05-02,DEBIT,2000.0,new deposit,RO09FBO
 741bfbdb-2a02-457e-a31b-4145db42d0db,2022-05-02,CREDIT,3000.0,loan granting,RO09FBOR77971890051360302664707462
 41aee598-860a-4c50-8f28-adbb3cc673ae,2022-05-02,DEBIT,500.0,money withdraw from ATM,RO09FBOR77971890051360302664707462
 a8c09597-b07c-4366-84fc-d095d34d143e,2022-05-02,DEBIT,0.0,ATM's withdraw fee,RO09FBOR77971890051360302664707462
+e1caf905-0526-4582-9f7e-24d37317edb6,2022-05-02,CREDIT,1000.0,transfer,RO09FBOR77971890051360302664707462
+7d96cd3f-9a67-4b6a-93ef-184d7934b185,2022-06-02,DEBIT,62.5,payment for loan,RO09FBOR77971890051360302664707462
+3878e1ca-a59b-4da9-b7cc-3cfe3ac4ec81,2022-06-02,DEBIT,25.0,interest for loan,RO09FBOR77971890051360302664707462
+13c8cc69-c9e2-47e0-8f52-4c42be5adae2,2022-07-02,DEBIT,62.5,payment for loan,RO09FBOR77971890051360302664707462
+d0537648-1d39-4894-b619-b514a6f2e509,2022-07-02,DEBIT,24.479166666666668,interest for loan,RO09FBOR77971890051360302664707462
+e381945a-4ddf-45d0-9ca8-21abd1222fa5,2022-08-02,CREDIT,2000.0,liquidate deposit,RO09FBOR77971890051360302664707462
+3217a57f-52b3-41ab-8476-9329056bf19d,2022-08-02,CREDIT,12.5,interest from deposit,RO09FBOR77971890051360302664707462
+03777d09-5ec3-4e6a-ace4-df3c2a74ed5c,2022-08-02,DEBIT,62.5,payment for loan,RO09FBOR77971890051360302664707462
+35a3736b-ec25-4b71-9f21-77f4d4b96d94,2022-08-02,DEBIT,23.958333333333332,interest for loan,RO09FBOR77971890051360302664707462
+9de644d1-62c0-480f-adca-8c47b54ff1d7,2022-09-02,DEBIT,62.5,payment for loan,RO09FBOR77971890051360302664707462
+e29970d0-c8a8-4b09-9ae1-375bee8fbb62,2022-09-02,DEBIT,23.4375,interest for loan,RO09FBOR77971890051360302664707462
+d10d3e7c-a3e7-4313-98a9-2128fc197c9b,2022-09-10,DEBIT,100.0,paid in exchage rate service,RO09FBOR77971890051360302664707462
+4c5a2cc3-fcbb-428b-947f-5b4cd8e33cf0,2022-09-10,CREDIT,20.263424518743665,received in exchange rate service,RO09FBOR64062931925950907483640704
+dca12e06-8896-4880-a68e-bf95301dd3ca,2022-05-02,CREDIT,1000.0,transfer,RO09FBOR54545018108522602821911997
+3e205af0-d0cc-45a7-a657-0e210a7e08b0,2022-05-02,CREDIT,1000.0,transfer,RO09FBOR99630222573054385015767051
+53f5d012-b09c-47ef-80eb-2032cf3708ef,2022-05-02,CREDIT,1000.0,transfer,RO09FBOR95564592353752061963174250
+3726f163-0b65-4617-83f8-631f7795cca7,2022-05-02,CREDIT,1000.0,transfer,RO09FBOR43248908097809710701388449
+bf13f240-c417-402d-8b7c-3add0b0f7253,2022-05-02,CREDIT,1000.0,transfer,RO09FBOR14604251942325876083032505
+832d10a5-1759-4c9e-8218-8f0591f27223,2022-05-02,CREDIT,1000.0,transfer,RO09FBOR29010589048258359095135357
+bafb8b03-04b1-42cc-a516-23eeaa686705,2022-05-02,CREDIT,1000.0,transfer,RO09FBOR76585660825878442506230487
